requirements.txt
langchain>=0.3.20,<1.0.0
langchain-community>=0.3.19
langchain-core>=0.3.41
langchain-google-genai>=0.0.11
langsmith>=0.1.125,<0.4
google-generativeai>=0.3.0
pypdf2>=3.0.1
python-dotenv>=1.0.0
pytesseract>=0.3.10
pypdfium2>=4.22.0
streamlit>=1.28.0
weaviate-client>=3.24.1